// Dynamics/Joints/b2DistanceJoint.d.ts
import { B2Vec2, B2Rot } from '../../Common/b2Math';
import { B2Joint, B2JointDef } from './b2Joint';
import { B2SolverData } from '../b2TimeStep';
import { B2Body } from '../b2Body';
export declare class B2DistanceJointDef extends B2JointDef {
    localAnchorA: B2Vec2;
    localAnchorB: B2Vec2;
    length: number;
    frequencyHz: number;
    dampingRatio: number;
    constructor();
    Initialize(b1: B2Body, B2: B2Body, anchor1: B2Vec2, anchor2: B2Vec2): void;
}
export declare class B2DistanceJoint extends B2Joint {
    m_frequencyHz: number;
    m_dampingRatio: number;
    m_bias: number;
    m_localAnchorA: B2Vec2;
    m_localAnchorB: B2Vec2;
    m_gamma: number;
    m_impulse: number;
    m_length: number;
    m_indexA: number;
    m_indexB: number;
    m_u: B2Vec2;
    m_rA: B2Vec2;
    m_rB: B2Vec2;
    m_localCenterA: B2Vec2;
    m_localCenterB: B2Vec2;
    m_invMassA: number;
    m_invMassB: number;
    m_invIA: number;
    m_invIB: number;
    m_mass: number;
    m_qA: B2Rot;
    m_qB: B2Rot;
    m_lalcA: B2Vec2;
    m_lalcB: B2Vec2;
    constructor(def: B2DistanceJointDef);
    GetAnchorA(out: B2Vec2): B2Vec2;
    GetAnchorB(out: B2Vec2): B2Vec2;
    GetReactionForce(inv_dt: number, out: B2Vec2): B2Vec2;
    GetReactionTorque(inv_dt: number): number;
    GetLocalAnchorA(): B2Vec2;
    GetLocalAnchorB(): B2Vec2;
    SetLength(length: number): void;
    Length(): number;
    SetFrequency(hz: number): void;
    GetFrequency(): number;
    SetDampingRatio(ratio: number): void;
    GetDampingRatio(): number;
    Dump(log: (format: string, ...args: any[]) => void): void;
    private static InitVelocityConstraints_s_P;
    InitVelocityConstraints(data: B2SolverData): void;
    private static SolveVelocityConstraints_s_vpA;
    private static SolveVelocityConstraints_s_vpB;
    private static SolveVelocityConstraints_s_P;
    SolveVelocityConstraints(data: B2SolverData): void;
    private static SolvePositionConstraints_s_P;
    SolvePositionConstraints(data: B2SolverData): boolean;
}